Revert "[Perf] Reference audio cache for MOSS-TTS-Local-v1.5 (LRU + single-flight dedup) (#743)"
This reverts commit 0c6f14729d0c78315468740a4dbd31e420a2a01b.
