+++ sglang_omni/preprocessing/cache_key.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import threading
-from collections import OrderedDict
 from pathlib import Path
 from typing import Any, Callable
 from urllib.parse import urlparse
@@ -50,114 +48,6 @@ def hash_file_sampled(
 
     payload = head + tail + f"|size:{file_size}".encode()
     return xxhash.xxh3_64(payload).hexdigest()
-
-
-_REF_PATH_HASH_MEMO_MAX_ITEMS = 1024
-_REF_PATH_HASH_SENTINEL_BYTES = 8192
-_REF_PATH_HASH_MEMO: OrderedDict[str, tuple[str, str]] = OrderedDict()
-_REF_PATH_HASH_MEMO_LOCK = threading.Lock()
-
-
-def _reference_path_hash_memo_key(path: Path) -> tuple[str, int] | None:
-    try:
-        if not path.is_file():
-            return None
-        stat_result = path.stat()
-        memo_key = (
-            f"{path.resolve()}:"
-            f"{stat_result.st_size}:"
-            f"{stat_result.st_mtime_ns}:"
-            f"{stat_result.st_ctime_ns}"
-        )
-        return memo_key, int(stat_result.st_size)
-    except OSError:
-        return None
-
-
-def _reference_path_sentinel(path: Path, file_size: int) -> str | None:
-    try:
-        chunk_size = min(_REF_PATH_HASH_SENTINEL_BYTES, file_size)
-        with path.open("rb") as f:
-            chunks = [f.read(chunk_size)]
-            if file_size > _REF_PATH_HASH_SENTINEL_BYTES:
-                middle_offset = max((file_size - chunk_size) // 2, 0)
-                f.seek(middle_offset)
-                chunks.append(f.read(chunk_size))
-            if file_size > 2 * _REF_PATH_HASH_SENTINEL_BYTES:
-                f.seek(max(file_size - chunk_size, 0))
-                chunks.append(f.read(chunk_size))
-        return hash_bytes(b"".join(chunks) + f"|size:{file_size}".encode())
-    except OSError:
-        return None
-
-
-def _get_reference_path_hash(memo_key: str, sentinel: str) -> str | None:
-    with _REF_PATH_HASH_MEMO_LOCK:
-        cached = _REF_PATH_HASH_MEMO.get(memo_key)
-        if cached is None:
-            return None
-        cached_sentinel, digest = cached
-        if cached_sentinel != sentinel:
-            _REF_PATH_HASH_MEMO.pop(memo_key, None)
-            return None
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        return digest
-
-
-def _get_reference_path_hash_by_memo_key(memo_key: str) -> str | None:
-    # Memo lookup ignoring the sentinel; trust_stat=True callers only.
-    with _REF_PATH_HASH_MEMO_LOCK:
-        cached = _REF_PATH_HASH_MEMO.get(memo_key)
-        if cached is None:
-            return None
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        return cached[1]
-
-
-def _put_reference_path_hash(memo_key: str, sentinel: str, digest: str) -> None:
-    with _REF_PATH_HASH_MEMO_LOCK:
-        _REF_PATH_HASH_MEMO[memo_key] = (sentinel, digest)
-        _REF_PATH_HASH_MEMO.move_to_end(memo_key)
-        while len(_REF_PATH_HASH_MEMO) > _REF_PATH_HASH_MEMO_MAX_ITEMS:
-            _REF_PATH_HASH_MEMO.popitem(last=False)
-
-
-def reference_path_cache_key(
-    path_like: str | Path, *, trust_stat: bool = False
-) -> str | None:
-    # Memoized full-content hash; the stat tuple skips rereads of stable files.
-    # Note(Jiaxin): trust_stat (opt-in, from #740) trusts the (size,mtime,ctime)
-    # stat tuple and skips the sentinel byte-read on memo hits; the accepted gap
-    # is same-size+mtime+ctime-with-different-content (reachable only by clock
-    # rollback). Default False keeps Higgs's sentinel path; keys are identical.
-    path = Path(str(path_like)).expanduser()
-    memo = _reference_path_hash_memo_key(path)
-    if memo is None:
-        return None
-    memo_key, file_size = memo
-
-    if trust_stat:
-        digest = _get_reference_path_hash_by_memo_key(memo_key)
-        if digest is not None:
-            return f"file:{digest}"
-
-    sentinel = _reference_path_sentinel(path, file_size)
-    if sentinel is None:
-        return None
-
-    if not trust_stat:
-        digest = _get_reference_path_hash(memo_key, sentinel)
-        if digest is not None:
-            return f"file:{digest}"
-
-    try:
-        digest = hash_bytes(path.read_bytes())
-    except OSError:
-        return None
-    if _reference_path_hash_memo_key(path) == memo:
-        # Always store the sentinel so default callers still validate this entry.
-        _put_reference_path_hash(memo_key, sentinel, digest)
-    return f"file:{digest}"
 
 
 def hash_media_item(item: Any) -> str | None:
